refactor: move data_vtack, make check_X_y without replacement
