+++ modAL/utils/combination.py
@@ -1,12 +1,4 @@
 import numpy as np
-import scipy.sparse as sp
-
-
-def data_vstack(blocks):
-    """Stack vertically both sparse and dense arrays."""
-    if sp.issparse(blocks[0]):
-        return sp.vstack(blocks)
-    return np.concatenate(blocks)
 
 
 def make_linear_combination(*functions, weights=None):
